pyproject.toml
[project]
name = "wearon-worker"
version = "0.1.0"
description = "WearOn Python worker (FastAPI + MediaPipe)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "pydantic>=2.10.0",
  "httpx>=0.28.0",
  "pillow>=11.0.0",
  "numpy>=2.1.0",
  "structlog>=24.4.0",
  "redis>=6.0.0",
  "mediapipe>=0.10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

